Refining

+++ src/main/java/com/example/taxcalculator/beans/Result.java
@@ -2,13 +2,15 @@ package com.example.taxcalculator.beans;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Result {
 
-    private Double grossAnnualSalary;
-    private Double grossMonthlySalary;
-    private Double netAnnualSalary;
-    private Double netMonthlySalary;
-    private Double annualTaxPaid;
-    private Double monthlyTaxPaid;
+    private BigDecimal grossAnnualSalary;
+    private BigDecimal grossMonthlySalary;
+    private BigDecimal netAnnualSalary;
+    private BigDecimal netMonthlySalary;
+    private BigDecimal annualTaxPaid;
+    private BigDecimal monthlyTaxPaid;
 }
